refactor: more UBID -> ZLDAP refactor
