+++ store/src/main/java/com/zimbra/cs/ldap/ZLdapFilter.java
@@ -5,23 +5,38 @@
 
 package com.zimbra.cs.ldap;
 
+import com.unboundid.ldap.sdk.Filter;
 import com.zimbra.cs.ldap.ZLdapFilterFactory.FilterId;
 
-public abstract class ZLdapFilter extends ZLdapElement {
-    
+public class ZLdapFilter extends ZLdapElement {
+
     private FilterId filterId;
-    
-    protected ZLdapFilter(FilterId filterId) {
+    private Filter filter;
+
+    public ZLdapFilter(FilterId filterId, Filter filter) {
         this.filterId = filterId;
+        this.filter = filter;
     }
-    
-    public abstract String toFilterString();
-    
+
+    @Override
+    public void debug() {
+    }
+
+    public Filter getNative() {
+        return filter;
+    }
+
     public FilterId getFilterId() {
         return filterId;
     }
-    
+
     public String getStatString() {
         return filterId.getStatString();
+    }
+
+    public String toFilterString() {
+        // cannot use this one, assertion values are all turned to lower case
+        // return getNative().toNormalizedString();
+        return getNative().toString();
     }
 }
