+++ store/src/main/java/com/zimbra/cs/ldap/ZSearchScope.java
@@ -5,33 +5,31 @@
 
 package com.zimbra.cs.ldap;
 
-public abstract class ZSearchScope {
+import com.unboundid.ldap.sdk.SearchScope;
+
+public class ZSearchScope {
+
     // only the entry specified by the base DN should be considered.
-    public static ZSearchScope SEARCH_SCOPE_BASE;
-    
-    // only entries that are immediate subordinates of the entry specified 
+    public static final ZSearchScope SEARCH_SCOPE_BASE = new ZSearchScope(SearchScope.BASE);
+
+    // only entries that are immediate subordinates of the entry specified
     // by the base DN (but not the base entry itself) should be considered.
-    public static ZSearchScope SEARCH_SCOPE_ONELEVEL;
-    
+    public static final ZSearchScope SEARCH_SCOPE_ONELEVEL = new ZSearchScope(SearchScope.ONE);
+
     // the base entry itself and any subordinate entries (to any depth) should be considered.
-    public static ZSearchScope SEARCH_SCOPE_SUBTREE;
-    
-    // any subordinate entries (to any depth) below the entry specified by the base DN should 
+    public static final ZSearchScope SEARCH_SCOPE_SUBTREE = new ZSearchScope(SearchScope.SUB);
+
+    // any subordinate entries (to any depth) below the entry specified by the base DN should
     // be considered, but the base entry itself should not be considered.
-    public static ZSearchScope SEARCH_SCOPE_CHILDREN;
-    
-    public abstract static class ZSearchScopeFactory {
-        protected abstract ZSearchScope getBaseSearchScope();
-        protected abstract ZSearchScope getOnelevelSearchScope();
-        protected abstract ZSearchScope getSubtreeSearchScope();
-        protected abstract ZSearchScope getChildrenSearchScope();
-    }
-    
-    public static void init(ZSearchScopeFactory factory) {
-        SEARCH_SCOPE_BASE = factory.getBaseSearchScope();
-        SEARCH_SCOPE_ONELEVEL = factory.getOnelevelSearchScope();
-        SEARCH_SCOPE_SUBTREE = factory.getSubtreeSearchScope();
-        SEARCH_SCOPE_CHILDREN = factory.getChildrenSearchScope();
+    public static final ZSearchScope SEARCH_SCOPE_CHILDREN = new ZSearchScope(SearchScope.SUBORDINATE_SUBTREE);
+
+    private final SearchScope searchScope;
+
+    private ZSearchScope(SearchScope searchScope) {
+        this.searchScope = searchScope;
     }
 
+    public SearchScope getNative() {
+        return searchScope;
+    }
 }
